add Directory UZ & IDS

+++ EFIDS/Abstract/IRepository.cs
@@ -12,11 +12,11 @@ namespace EFIDS.Abstract
         Database Database { get; }
         IQueryable<T> Context { get; }
         IEnumerable<T> Get();       // получение всех объектов
-        T Get(long id);              // получение одного объекта по id
+        T Get(int id);              // получение одного объекта по id
         void Add(T item);           // создание объекта
         void Update(T item);        // обновление объекта
         void AddOrUpdate(T item);   // добавить или обновить
-        void Delete(long id);        // удаление объекта по id
+        void Delete(int id);        // удаление объекта по id
         int Save();                 // сохранение изменений в базе
         T Refresh(T item);          // Обновить объект с базаой
     }
